Only calculate local eslint exec once
Rather than finding the ESLint executable each time I load a JS file,
change it so it's done the first time I touch a js file and then not
again. Because I never use Vim to jump between projects, I can be
certain that once I find a local eslint executable to use I'll be using
it for the duration of that Vim's lifetime.